backend/requirements.txt
fastapi
uvicorn
pypdf
faiss-cpu
numpy
google-genai
python-dotenv
python-multipart